Make separate physics layers for red / blue enemy.
They will always exist on the layer enemy and the layour correct to their colour, but monitor for collisions only on the mask correct to their colour.

Therefore projectiles and other objects can detect them on the layer enemy AND the enemyred layer / enemyblue layer. But they are blind to the enemy mask and ONLY monitor on the enemyblue / enemyred layer.